company-ai-chatbot/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==1.4.49
python-dotenv==1.0.0
asyncpg==0.29.0
httpx==0.25.2
pydantic==2.5.0
